safari-600.7-branch/LayoutTests/platform/mac/fast/events/non-roman-key-code-expected.txt

Test that non-Roman keys get proper non-null key codes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.charCode is 0
PASS event.keyCode is 70
PASS successfullyParsed is true

TEST COMPLETE
